# contracts/operator.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Max operator API",
    "version": "1.0.0",
...
  },
  "servers": [{ "url": "https://{host}/operator/v1", "variables": { "host": { "default": "api.example.com", "description": "This deployment's backend host" } } }],
  "security": [{ "serviceIdentity": [] }],
...
  "paths": {
...
    "/billing/quotes": {
      "get": {
        "tags": ["billing"],
        "operationId": "listQuotes",
        "summary": "All offline-payment quotes, with the user's e-mail",
        "description": "Provider field names are mapped for the console: buyer.taxId → buyer.oib, note → paidNote; userEmail is added by the core. The provider's status and JSON body are relayed unchanged (contracts/billing-provider.openapi.json); a non-JSON error answer becomes 502.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Forwarded to the provider (creating | issued | paid | expired | cancelled | failed)"
          }
        ],
        "responses": {
          "200": {
            "description": "Quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": { "type": ["object", "null"], "additionalProperties": true, "description": "Provider configuration as reported by it" },
                    "quotes": { "type": "array", "items": { "$ref": "#/components/schemas/AdminQuote" } }
                  },
                  "required": ["config", "quotes"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "502": {
            "description": "BILLING_PROVIDER_BAD_ANSWER — the provider answered an error status with a non-JSON body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      },
      "post": {
        "tags": ["billing"],
        "operationId": "createQuote",
        "summary": "Issue a quote for a registered user",
        "description": "The core prices the plan from its own catalogue (EUR, tax-exclusive, recurring) and fills the buyer country from the user's profile. Refused while the user has an active card subscription. On success the provider's answer (typically 201 {quote}) is relayed in the provider's own field names (buyer.taxId, note), not the mapped list shape. Audited as billing.quote.create. The provider's status and JSON body are relayed unchanged (contracts/billing-provider.openapi.json); a non-JSON error answer becomes 502.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string", "format": "uuid", "description": "A registered user" },
                  "plan": { "type": "string", "description": "Plan key of this deployment's catalogue" },
                  "seats": { "type": "integer", "description": "Per-seat plans only; clamped to the plan's minimum … 1000" },
                  "company": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string", "maxLength": 200 },
                      "oib": { "type": "string", "maxLength": 40, "description": "Buyer tax id" },
                      "street": { "type": "string", "maxLength": 200 },
                      "postalCode": { "type": "string", "maxLength": 20 },
                      "city": { "type": "string", "maxLength": 120 }
                    },
                    "description": "Strings trimmed and truncated; missing → \"\""
                  },
                  "email": { "type": "string", "description": "Recipient; blank → the user's e-mail" },
                  "provisional": { "type": "boolean", "default": false, "description": "Ask for provisional access until payment" },
                  "sendEmail": { "type": "boolean", "default": true }
                },
                "required": ["userId", "plan"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issued (relayed)",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "quote": { "type": "object", "additionalProperties": true } } } } }
          },
          "400": {
            "description": "USER_REQUIRED | PLAN_INVALID — or a relayed provider validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "USER_NOT_FOUND", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } } },
          "409": {
            "description": "STRIPE_SUBSCRIPTION_ACTIVE — or a relayed provider conflict",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "502": {
            "description": "BILLING_PROVIDER_BAD_ANSWER — the provider answered an error status with a non-JSON body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      }
    },
    "/billing/users/{userId}/billing": {
      "get": {
        "tags": ["billing"],
        "operationId": "quotePrefill",
        "summary": "Prefill for a new quote: stored buyer data and eligibility",
        "description": "A malformed or unknown userId answers 404 (not 400 as on the data routes).",
        "parameters": [{ "$ref": "#/components/parameters/actor" }, { "$ref": "#/components/parameters/userId" }],
        "responses": {
          "200": {
            "description": "Prefill",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": ["string", "null"] },
                    "buyer": { "$ref": "#/components/schemas/AdminBuyer" },
                    "complete": { "type": "boolean", "description": "All buyer fields present" },
                    "stripeActive": { "type": "boolean", "description": "An active card subscription blocks a quote" },
                    "provisionalAvailable": { "type": "boolean" },
                    "country": { "type": ["string", "null"] },
                    "countrySupported": { "type": "boolean", "description": "False only when the provider says so" }
                  },
                  "required": ["email", "buyer", "complete", "stripeActive", "provisionalAvailable", "country", "countrySupported"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "USER_NOT_FOUND", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } } },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      }
    },
    "/billing/quotes/{id}/{action}": {
      "post": {
        "tags": ["billing"],
        "operationId": "quoteAction",
        "summary": "Act on a quote: mark-paid | resend | cancel | retry",
        "description": "The body is forwarded unchanged: mark-paid {paidAt?, note?} | resend {email?} | cancel {note?} | retry {}. The provider's answer (typically 200 {quote}) is relayed in its own field names. Audited as billing.quote.<action with _>. The provider's status and JSON body are relayed unchanged (contracts/billing-provider.openapi.json); a non-JSON error answer becomes 502.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "action", "in": "path", "required": true, "schema": { "type": "string", "enum": ["mark-paid", "resend", "cancel", "retry"] } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paidAt": { "type": "string", "description": "mark-paid" },
                  "note": { "type": "string", "description": "mark-paid, cancel" },
                  "email": { "type": "string", "description": "resend" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated (relayed)",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "quote": { "type": "object", "additionalProperties": true } } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "NOT_FOUND (malformed id or unknown action) — or a relayed provider 404",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "409": { "description": "Relayed provider conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } } },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "502": {
            "description": "BILLING_PROVIDER_BAD_ANSWER — the provider answered an error status with a non-JSON body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      }
    },
    "/billing/quotes/{id}/pdf": {
      "get": {
        "tags": ["billing"],
        "operationId": "quotePdf",
        "summary": "A quote's PDF",
        "description": "Relayed as application/pdf (Content-Disposition passed through, Cache-Control: private, no-store); provider errors as JSON.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "PDF", "content": { "application/pdf": { "schema": { "type": "string", "format": "binary" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "QUOTE_NOT_FOUND (malformed id) — or a relayed provider 404",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "502": {
            "description": "BILLING_PROVIDER_BAD_ANSWER — the provider answered an error status with a non-JSON body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      }
    },
    "/billing/invoices/{docId}/pdf": {
      "get": {
        "tags": ["billing"],
        "operationId": "invoicePdf",
        "summary": "An invoice's PDF",
        "description": "Relayed as application/pdf (Content-Disposition passed through, Cache-Control: private, no-store); provider errors as JSON.",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "name": "docId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[0-9A-Za-z:_-]{1,64}$" } }
        ],
        "responses": {
          "200": { "description": "PDF", "content": { "application/pdf": { "schema": { "type": "string", "format": "binary" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "INVOICE_NOT_FOUND (malformed id) — or a relayed provider 404",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "500": { "$ref": "#/components/responses/BillingInternal" },
          "502": {
            "description": "BILLING_PROVIDER_BAD_ANSWER — the provider answered an error status with a non-JSON body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          },
          "503": {
            "description": "BILLING_PROVIDER_UNAVAILABLE — seam not configured (BILLING_PROVIDER_URL + BILLING_SERVICE_SECRET) or the provider is unreachable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceIdentity": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google OIDC ID token",
        "description": "A Google-signed OIDC identity token of a service account (no shared password). Accepted when the signature verifies against Google's keys, email_verified is true and the email (case-insensitive) is listed in OPERATOR_API_CALLERS (separated by ',', ';' or whitespace — ';' survives gcloud's comma-split env flags), and aud is one of OPERATOR_API_AUDIENCE (same separators, trailing '/' ignored) or, when that is unset, this host's origin https://{Host} or https://{Host}/operator/v1. OPERATOR_API_CALLERS unset or empty ⇒ the API does not exist: every path answers 404 before the token is looked at. Outside production (NODE_ENV ≠ production) the static OPERATOR_API_DEV_TOKEN is accepted as well (audit actor fallback 'dev'), still only when OPERATOR_API_CALLERS is set."
      }
    },
    "parameters": {
      "actor": {
        "name": "X-Operator-Actor",
        "in": "header",
        "required": false,
        "description": "The human or tool behind this request (e.g. an operator's e-mail). Trimmed; anything not matching the pattern is ignored silently. Recorded as admin_audit.actor for every audited action and echoed by GET /whoami; absent or invalid → the caller's service-account e-mail ('dev' with the dev token). Trusted only because the caller itself is authenticated.",
        "schema": { "type": "string", "pattern": "^[\\w.@:+-]{1,100}$" }
      },
      "userId": { "name": "userId", "in": "path", "required": true, "description": "public.users id", "schema": { "type": "string", "format": "uuid" } },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing bearer token | Invalid caller token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "BillingInternal": { "description": "INTERNAL — unexpected failure in the core", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillingError" } } } }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["detail"],
        "properties": { "detail": { "type": "string" }, "error": { "type": "string", "description": "Exception message (global error handler only)" } },
        "additionalProperties": true
      },
      "BillingError": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "e.g. BILLING_PROVIDER_UNAVAILABLE, BILLING_PROVIDER_BAD_ANSWER, USER_REQUIRED, USER_NOT_FOUND, STRIPE_SUBSCRIPTION_ACTIVE, PLAN_INVALID, NOT_FOUND, QUOTE_NOT_FOUND, INVOICE_NOT_FOUND, INTERNAL, or a relayed provider code"
          },
          "detail": { "type": "string" },
          "field": { "type": "string", "description": "Relayed provider validation errors only" }
        },
        "additionalProperties": true
      },
...
      "AdminBuyer": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "oib": { "type": "string", "description": "Buyer tax id (provider: taxId)" },
          "street": { "type": "string" },
          "postalCode": { "type": "string" },
          "city": { "type": "string" }
        }
      },
      "AdminQuote": {
        "type": "object",
        "additionalProperties": true,
        "description": "The provider's AdminQuote (contracts/billing-provider.openapi.json) with buyer mapped to AdminBuyer, note renamed paidNote, and userEmail added by the core.",
        "properties": {
          "id": { "type": "string" },
          "number": { "type": ["string", "null"] },
          "status": { "type": "string", "enum": ["creating", "issued", "paid", "expired", "cancelled", "failed"] },
          "plan": { "type": "string" },
          "planName": { "type": "string" },
          "seats": { "type": "integer" },
          "netAmount": { "type": "number" },
          "grossAmount": { "type": "number" },
          "currency": { "type": "string" },
          "periodMonths": { "type": "integer" },
          "validUntil": { "type": ["string", "null"] },
          "createdAt": { "type": "string" },
          "createdBy": { "type": "string", "enum": ["user", "admin"] },
          "recipientEmail": { "type": "string" },
          "emailStatus": { "type": "string", "enum": ["pending", "sent", "failed", "skipped"] },
          "emailSentAt": { "type": ["string", "null"] },
          "paymentReference": { "type": ["string", "null"] },
          "provisionalUntil": { "type": ["string", "null"] },
          "paidAt": { "type": ["string", "null"] },
          "paidSource": { "type": ["string", "null"] },
          "accessUntil": { "type": ["string", "null"] },
          "invoiceNumbers": { "type": "array", "items": { "type": "string" } },
          "canDownload": { "type": "boolean" },
          "userId": { "type": "string" },
          "userEmail": { "type": ["string", "null"] },
          "buyer": { "$ref": "#/components/schemas/AdminBuyer" },
          "paidNote": { "type": ["string", "null"] },
          "paidAmount": { "type": ["number", "null"] },
          "lastError": { "type": ["string", "null"] }
        }
      }
    }
  }
}